+++ doc/release-notes.md
@@ -94,6 +94,9 @@ Build System
 Updated settings
 ----------------
 
+- The `-debug=db` logging category, which was deprecated in 0.20 and replaced by
+  `-debug=walletdb` to distinguish it from `coindb`, has been removed. (#19202)
+
 Changes to Wallet or GUI related settings can be found in the GUI or Wallet  section below.
 
 New settings